+++ mirrord/layer/src/detour.rs
@@ -172,10 +172,6 @@ pub(crate) enum Bypass {
     #[cfg(target_os = "macos")]
     ExecOnNonExistingFile(String),
 
-    /// No args where changed in `intercept_tmp_dir`.
-    #[cfg(target_os = "macos")]
-    NoTempDirInArgv,
-
     /// Reached `MAX_ARGC` while running
     /// `intercept_tmp_dir`
     #[cfg(target_os = "macos")]
@@ -289,12 +285,17 @@ impl<S> Detour<S> {
         }
     }
 
-    /// Like Result's `ok()`.
-    #[cfg(target_os = "macos")] // Currently only used in macos, remove to use in linux code.
-    pub(crate) fn success(self) -> Option<S> {
+    /// Return the contained `Success` value or a provided default if `Bypass` or `Error`.
+    ///
+    /// To be used in hooks that are deemed non-essential, and the run should continue even if they
+    /// fail.
+    /// Currently defined only on macos because it is only used in macos-only code.
+    /// Remove the cfg attribute to enable using in other code.
+    #[cfg(target_os = "macos")]
+    pub(crate) fn unwrap_or(self, default: S) -> S {
         match self {
-            Detour::Success(s) => Some(s),
-            _ => None,
+            Detour::Success(s) => s,
+            _ => default,
         }
     }
 }