+++ mirrord/operator/src/client.rs
@@ -118,7 +118,7 @@ where
                     if let Some(client_message) = client_message {
                         if let Ok(payload) = bincode::encode_to_vec(client_message, bincode::config::standard()) {
                             if let Err(err) = connection.send(payload.into()).await {
-                                eprintln!("{:?}", err);
+                                eprintln!("{err:?}");
 
                                 break;
                             }
@@ -133,7 +133,7 @@ where
                     if let Some(Ok(Message::Binary(payload))) = daemon_message {
                         if let Ok((daemon_message, _)) = bincode::decode_from_slice::<DaemonMessage, _>(&payload, bincode::config::standard()) {
                             if let Err(err) = daemon_tx.send(daemon_message).await {
-                                eprintln!("{:?}", err);
+                                eprintln!("{err:?}");
 
                                 break;
                             }